demonstration of recurrence across string graph nodes
As of this commit, issuing the following command will demonstrate the ability
to seed an alignment with the H and E vectors from a previous alignment.

    ./ssw_example ATGC CTGA ATGCCTGA

This equates to a string graph of the form:

    ATGC -> CTGA

And we are aligning against a string that matches the entire string graph.

Several features which will be required for use in partial order / string
alignment are still missing.

In the case that a sequence node has multiple predecessors, we need to generate
the H vector as the maximum of inbound H vectors on a per-cell basis.  The
traceback can independently determine the direction taken by the optimal
alignment across the node boundaries.

The same will need to be done for the E vector.  On face value, this seems a
bit more complex and I am still nivestigating how this should be done.