avoid traceback error by correctly detecting when we are using uint16_t scores
This caused major woes downstream, but thankfully is only likely in the case of
longer reads. I detected the problem on 151 bp XTen reads, which apparently
were triggering the word-sized scoring scheme, but an older check for score
size was based on the maximum score, yielding problems when 16-bit scores were
used but the score didn't pass 255. This was a common scenario with long reads.